requirements.txt
gunicorn
Flask
Flask-SQLAlchemy
Flask-Script
Flask-Migrate
Flask-WTF
Flask-Login
Flask-Mail
Flask-Celery-Helper
Flask-Babel
Flask-Elasticsearch
Flask-Failsafe
Flask-Babel
translitcodec
pytz
passlib
bcrypt
requests
pymysql
openpyxl
cryptography
qrcode
celery
elasticsearch